% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.MBOExampleRunMultiCrit}
\alias{autoplot.MBOExampleRunMultiCrit}
\title{Plot example run, either in 1D or 2D.}
\usage{
autoplot.MBOExampleRunMultiCrit(object, iters, pause = TRUE, y1lim = NULL,
  y2lim = NULL, ...)
}
\arguments{
\item{object}{[\code{}]\cr
Object of type \code{MBOExampleRunMultiCrit}.}

\item{iters}{[\code{integer}]\cr
Selected iterations of \code{x} to display.
Default is all iterations.}

\item{pause}{[\code{logical(1)}]\cr
Pause after each iteration? In this case the plots are arranged in a grid making use
of the gridExtra package.
Default is \code{TRUE}.}

\item{y1lim}{[\code{numeric(2)}]\cr
Axis limits for the first target value.
Default is range of y1-values evaluated in run object \code{y1}.}

\item{y2lim}{[\code{numeric(2)}]\cr
Axis limits for the second target value.
Default is range of y2-values evaluated in run object \code{y2}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{list}]. List of lists. For each iteration the sublist contains separate ggplot
  objects for the pareto set and the pareto front.
}
\description{
Plot example run, either in 1D or 2D.
}

