% Generated by roxygen2 (4.0.1): do not edit by hand
\name{autoplot.ParEGOExampleRun}
\alias{autoplot.ParEGOExampleRun}
\title{Plot example run, either in 1D or 2D.}
\usage{
autoplot.ParEGOExampleRun(run, iters, pause = TRUE, y1lim = NULL,
  y2lim = NULL, ...)
}
\arguments{
\item{run}{[\code{}]\cr
Objective function.}

\item{iters}{[\code{integer}]\cr
Selected iterations of \code{x} to display.
Default is all iterations.}

\item{pause}{[\code{logical(1)}]\cr
Pause after each iteration?
Default is \code{TRUE}.}

\item{y1lim}{[\code{numeric(2)}]\cr
axis limits for the Plot
Default is range of y1-values evaluated in run object \code{y1}.}

\item{y2lim}{[\code{numeric(2)}]\cr
axis limits for the Plot
Default is range of y2-values evaluated in run object \code{y2}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{list}]. List containing seperate ggplot plots for each iteration.
}
\description{
FIXME: Describe Plot
FIXME: This function is not well tested - not sure if we want to export it now
}

