% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMBO.R
\name{finalizeSMBO}
\alias{finalizeSMBO}
\title{Finalizes the SMBO Optimization}
\usage{
finalizeSMBO(opt.state)
}
\arguments{
\item{opt.state}{[\code{\link{OptState}}]
The optimization state.}
}
\value{
[\code{\link{MBOSingleObjResult}} | \code{\link{MBOMultiObjResult}}]
}
\description{
Returns the common mlrMBO result object.
}
