% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getInfillCrits}
\alias{getInfillCrits}
\title{Returns the names of all currently supported infill criteria.}
\usage{
getInfillCrits()
}
\value{
[\code{character}]
}
\description{
Returns the names of all currently supported infill criteria.
}

