% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMBOInfillCrit.R
\name{getMBOInfillCrit}
\alias{getMBOInfillCrit}
\alias{getMBOInfillCritParams}
\alias{getMBOInfillCritParam}
\alias{getMBOInfillCritName}
\alias{getMBOInfillCritId}
\alias{hasRequiresInfillCritStandardError}
\alias{getMBOInfillCritComponents}
\title{Get properties of MBO infill criterion.}
\usage{
getMBOInfillCritParams(x)

getMBOInfillCritParam(x, par.name)

getMBOInfillCritName(x)

getMBOInfillCritId(x)

hasRequiresInfillCritStandardError(x)

getMBOInfillCritComponents(x)
}
\arguments{
\item{x}{[\code{\link{MBOInfillCrit}}]\cr
Infill criterion.}

\item{par.name}{[\code{character(1)}]\cr
Parameter name.}
}
\description{
Returns properties of an infill criterion, e.g., name or id.
}
