% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSupportedInfillOptFunctions}
\alias{getSupportedInfillOptFunctions}
\title{Get supported infill-criteria optimizers.}
\usage{
getSupportedInfillOptFunctions()
}
\value{
[\code{list}] List of supported infill optimizers.
}
\description{
Get supported infill-criteria optimizers.
}

