% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSupportedMultipointInfillOptFunctions}
\alias{getSupportedMultipointInfillOptFunctions}
\title{Get supported multipoint infill-criteria opimizers.}
\usage{
getSupportedMultipointInfillOptFunctions()
}
\value{
[\code{list}] List of supported multipoint infill optimizers.
}
\description{
Get supported multipoint infill-criteria opimizers.
}

