% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infill_crits_interface.R
\name{infillcrits}
\alias{infillcrits}
\alias{makeMBOInfillCriterionAEI}
\alias{makeMBOInfillCriterionCB}
\alias{makeMBOInfillCriterionDIB}
\alias{makeMBOInfillCriterionEI}
\alias{makeMBOInfillCriterionEQI}
\alias{makeMBOInfillCriterionMeanResponse}
\alias{makeMBOInfillCriterionStandardError}
\title{Infill criteria.}
\usage{
makeMBOInfillCriterionMeanResponse()

makeMBOInfillCriterionStandardError()

makeMBOInfillCriterionEI()

makeMBOInfillCriterionCB(cb.lambda = 1, cb.inflate.se = FALSE,
  cb.pi = NULL)

makeMBOInfillCriterionAEI(aei.use.nugget = FALSE)

makeMBOInfillCriterionEQI(eqi.beta = 0.75)

makeMBOInfillCriterionDIB(cb.lambda = 1, cb.inflate.se = FALSE,
  cb.pi = NULL)
}
\arguments{
\item{cb.lambda}{[\code{numeric(1)}]\cr
Lambda parameter for confidence bound infill criterion.
Only used if \code{crit == "cb"}, ignored otherwise.
Default is 1.}

\item{cb.inflate.se}{[\code{logical(1)}]\cr
Try to inflate or deflate the estimated standard error to get to the same scale as the mean?
Calculates the range of the mean and standard error and multiplies the standard error
with the quotient of theses ranges.
Default is \code{FALSE}.}

\item{cb.pi}{[\code{numeric(1)}]\cr
Probability-of-improvement value to determine the lambda parameter for cb infill criterion.
It is an alternative to set the trade-off between \dQuote{mean} and \dQuote{se}.
Only used if \code{crit == "cb"}, ignored otherwise.
If specified, \code{cb.lambda == NULL} must hold.
Default is \code{NULL}.}

\item{aei.use.nugget}{[\code{logical(1)}]\cr
Should the nugget effect be used for the pure variance estimation for augmented
expected improvement?
Default is \code{FALSE}.}

\item{eqi.beta}{[\code{numeric(1)}]\cr
Beta parameter for expected quantile improvement criterion.
Default is 0.75.}
}
\description{
\pkg{mlrMBO} contains most of the most popular infill criteria.
}

