% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsynWrapper.R
\name{makeAsynWrapper}
\alias{makeAsynWrapper}
\title{Fuse learner with imputation of y}
\usage{
makeAsynWrapper(learner, aw.quantiles = NULL, aw.mc.iters = NULL)
}
\arguments{
\item{learner}{[\code{Learner}]\cr
mlr Learner}

\item{aw.quantiles}{[\code{numeric}]\cr
Quantiles that should be calculated for imputation.
Default is \code{NULL}.}

\item{aw.mc.iters}{[\code{integer(1)}]\cr
Number of samples for the monte carlo method to draw from the kriging model to calculate the EEI.
Default is \code{NULL}.
Only one of the above parameters can be given.}
}
\value{
Wrapped Learner
}
\description{
This wrapper will train the underling learner with the set of the task that has no NAs in y.
For the missing y values we will predict a set of quantiles.
All combinations of those quantile values will be taken to replace the missing values leading to multiple different new tasks.
We then will train a model for each of those tasks.
}
\seealso{
Other wrapper: \code{\link{makeDiscreteWrapper}}
}
