% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteWrapper.R
\name{makeDiscreteWrapper}
\alias{makeDiscreteWrapper}
\title{Treats a categorical variable as a seperator}
\usage{
makeDiscreteWrapper(learner, discw.category = NULL, discw.na.val = NULL)
}
\arguments{
\item{learner}{[\code{Learner}]\cr
mlr Learner}

\item{discw.category}{[\code{charachter(1)}]\cr
name of the column that indicates the categorical value to split before training.
If value is NULL (default) then the first categorical value will be taken into account.}

\item{discw.na.val}{[\code{numeric(1)}]\cr
Number that structural missing values should be imputed with.}
}
\value{
Wrapped Learner
}
\description{
This wrapper will make sure that the learner sees numerical values as independent in combination with each categorical variable
}
\seealso{
Other wrapper: \code{\link{makeAsynWrapper}}
}

