% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infill_crits_interface.R
\name{makeMBOInfillCriterion}
\alias{makeMBOInfillCriterion}
\title{Create an infill critertion.}
\usage{
makeMBOInfillCriterion(infill.fun, name, id, minimize = TRUE,
  params = list())
}
\arguments{
\item{name}{[\character{1}]\cr
Full name of the criterion.}

\item{id}{[\character{1}]\cr
Short name of the criterion.
Used internally and in plots.}

\item{minimize}{[\character(1)]\cr
Shall the criterion be minimized or maximized?
Default is \code{TRUE}.}

\item{params}{[\code{list}]\cr
Name list of parameters. Infill criteria might depend on parameters.
These parameters should be stored in the \code{params} list in order
to access them inside mlrMBO.
Default is the empty list.}
}
\value{
[\code{MBOInfillCriterion}]
  Basically \code{infill.fun} with an additional class.
  All other parameters are appended as attributes.
}
\description{
The infill criterion guides the model based search process.
The most prominent infill criteria, e.g., expected improvement, lower
confidence bound and others, are already implemented in mlrMBO. Moreover,
the package allows for the creation of proprietary infill criteria.
}

