% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMboKrigingLearner.R
\name{makeMboKrigingLearner}
\alias{makeMboKrigingLearner}
\title{Automatically generate the right Kriging learner for the given problem}
\usage{
makeMboKrigingLearner(control, fun, ...)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{fun}{[\code{smoof_function}] \cr
The same objective function which is also passed to \code{\link{mbo}}.}

\item{...}{[any]\cr
Parameters passed to \code{\link[DiceKriging]{km}}.
Will overwrite mlrMBOs recomendations.}
}
\value{
[\code{Learner}]
}
\description{
Based on in the function is noisy or not this function generates a kriging learner whichs settings are the right ones for mbo.
Additionally it determines the nugget for each model fit based on the variance of the target function.
}

