% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeMBOMultiFidControl}
\alias{makeMBOMultiFidControl}
\title{Generates the control object for the multiFid infill criteria}
\usage{
makeMBOMultiFidControl(fid.param, cor.grid.points = 10L, costs = NULL,
  force.last.level.evals = 10L)
}
\arguments{
\item{fid.param}{[\code{character(1)}]\cr
The name of the parameter which increases the performance but also calculation costs. Has to belong to a discrete Parameter.}

\item{cor.grid.points}{[\code{integer(1)}]\cr
Numbers of points used to calculate the correlation between the different levels of the \code{multiFid.fid.param}.}

\item{costs}{[\code{function}]\cr
Vektorized (?) cost function with the params \code{cur} and \code{last}.}

\item{force.last.level.evals}{[\code{integer(1)}]
How many evaluations should be done on the last value of fid.param.}
}
\value{
\code{MBOMultiFidControl}
}
\description{
Generates the control object for the multiFid infill criteria
}

