% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTrafoFunction}
\alias{makeTrafoFunction}
\title{Create transformation function for MBOExampleRun.}
\usage{
makeTrafoFunction(name, fun)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of the transformation.}

\item{fun}{[\code{function}]\cr
R function which expects a numeric vector.}
}
\value{
Object of type MBOTrafoFunction.
}
\description{
Create transformation function for MBOExampleRun.
}
\seealso{
\link{trafos}
}

