% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLearner.R
\name{mbo_default_learner}
\alias{mbo_default_learner}
\title{Default learner.}
\description{
If only numerical and/or integer parameters are present, we use
\dQuote{regr.km}, which is kriging / a gaussian process from package DiceKriging.
If the objective is noisy, we \code{nugget.estim} is set to \code{TRUE}.
If the objective is deterministic, we set nugget = 10^-3 for numerical stabiloity.


depending on whether we have noisy observations or not.
If a least one parameter is discrete, we use \dQuote{regr.randomForest}.
}

