% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_mbo_OptPath.R
\name{mbo_OptPath}
\alias{mbo_OptPath}
\title{OptPath in mlrMBO}
\description{
In mlrMBO the \code{\link[ParamHelpers]{OptPath}} contains extra information next to the information documented in \code{\link[ParamHelpers]{OptPath}}].

The extras are:
\describe{
\item{train.time}{Time to train the model(s) that produced the points. Only the first slot of the vector is used (if we have multiple points), rest are NA.}
\item{propose.time}{Time needed to propose the point. If we have indivual timings from the proposal mechanism, we have one differrent value per point here. If all were generated in one go, we only have one timing, we store it in the slot for the first point, rest are NA.}
\item{errors.model}{Possible Error Messages. If point-producing model(s) crashed they are replicated for all n points, if only one error message was passed we store it for the first point, rest are NA.}
\item{filter.replace}{If point filtering was enabled, we store whether the point was replaced by a random one.}
\item{multipoint.cb.lambdas}{Random lambda-value used in q-CB point proposal. One lambda for each point in that case.}
\item{.weight}{Weight vector sampled for multipoint ParEGO}
}
}

