% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboContinue.R
\name{mboContinue}
\alias{mboContinue}
\title{Continues an mbo run from a save-file.}
\usage{
mboContinue(file)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
File path of saved MBO state.
See \code{save.on.disk.at} argument of \code{\link{MBOControl}} object.}
}
\value{
See \code{\link{mbo}}.
}
\description{
Useful if your optimization is likely to crash,
so you can continue from a save point and will not lose too much information and runtime.
}
