% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboFinalize.R
\name{mboFinalize}
\alias{mboFinalize}
\title{Finalizes an mbo run from a save-file.}
\usage{
mboFinalize(file)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
File path of saved MBO state.
See \code{save.on.disk.at} argument of \code{\link{MBOControl}} object.}
}
\value{
See \code{\link{mbo}}.
}
\description{
Useful if your optimization didn't terminate but you want a results nonetheless.
}
