% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MBOSingleObjResult}
\alias{MBOSingleObjResult}
\title{Single-Objective result object.}
\description{
\itemize{
  \item{x [\code{list}]}{Named list of proposed optimal parameters.}
  \item{y [\code{numeric(1)}]}{Value of objective function at \code{x},
    either from evals during optimization or from requested final evaluations,
    if those were greater than 0.}
  \item{opt.path [\code{\link[ParamHelpers]{OptPath}}]}{Optimization path.
    Includes all evaluated points and additional information.
    You can convert it via \code{as.data.frame}.}
  \item{models [List of \code{\link[mlr]{WrappedModel}}]}{List of saved regression models.}
}
}

