% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptProblem.R
\name{OptProblem}
\alias{OptProblem}
\title{OptProblem object.}
\description{
The OptProblem contains all the constants values which define a OptProblem within our MBO Steps.
It is an environment and is always pointed at by the OptState.
}
