% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptResult.R
\name{OptResult}
\alias{OptResult}
\title{OptResult object.}
\description{
The OptResult stores all entities which are not needed while optimizing but are needed to build the final result.
It can contains fitted surrogate models at certain times as well as resample objects.
When the optimization ended it will contain the [\code{MBOResult}].
}
