% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptState.R
\name{OptState}
\alias{OptState}
\title{OptState object.}
\description{
The OptState is the central component of the mbo iterations.
This enviroment contains every necessary information needed during optimization in MBO.
It also links to the \code{\link{OptProblem}} and to the \code{\link{OptResult}}.
}
