% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.MBOMultiObjResult}
\alias{plot.MBOMultiObjResult}
\title{Plot method for multi-objective result objects.}
\usage{
\method{plot}{MBOMultiObjResult}(x, infill.crit, iters = NULL, alpha = TRUE,
  pause = TRUE, log.infill.crit = FALSE, ...)
}
\arguments{
\item{x}{[\code{MBOMultiObjResult}]\cr
Multi-Objective mlrMBO result object.}

\item{infill.crit}{[\code{character(1)}]\cr
Infill criterion used for optimization.}

\item{iters}{[\code{numeric} | NULL]\cr
Vector of iteration which should be plotted one after another. If \code{NULL},
which is the default, all iterations are plotted.}

\item{alpha}{[\code{logical(1)}]\cr
Activates or deactivates the alpha fading for the parallel X-space plot. Default is \code{TRUE}.}

\item{pause}{[\code{logical(1)}]\cr
Pause after each iteration?
Default is \code{TRUE}.}

\item{log.infill.crit}{[\code{logical(1)}]\cr
Should the infill criterion be log-transformed? Default is \code{FALSE}.}

\item{...}{[\code{list}]\cr
Not used.}
}
\value{
Nothing
}
\description{
Plot method for multi-objective result objects.
}

