% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OptState.R
\name{plot.OptState}
\alias{plot.OptState}
\title{Generate ggplot2 Object}
\usage{
\method{plot}{OptState}(x, ...)
}
\arguments{
\item{x}{[\code{OptState}]
The OptState.}

\item{...}{[any] \cr
Not used.}
}
\description{
Plots the values of the infill criterion for a 1- and 2-dimensional numerical search space for a given \code{\link{OptState}}.
}
