% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot for Learner Predictions}
\usage{
plot_prediction(learner, task, grid_points = 100L, expand_range = 0)
}
\arguments{
\item{expand_range}{(`numeric(1)`)\cr
Expand the prediction range over the range of the task.}

\item{Learner}{([mlr3::Learner])}

\item{Task}{([mlr3::Task])}

\item{grid_ponints}{(`integer(1)`)\cr
Resolution of the grid.}
}
\value{
[ggplot2::ggplot()] object.
}
\description{
Generates plots for [mlr3::Learner], and [mlr3::Task]
}
\examples{
library(mlr3)
library(mlr3viz)

task = tsk("spam")
learner = lrn("classif.rpart", predict_type = "prob")
plot_prediction(learner, task)
}
