% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_MBOResult.R
\name{plotMBOResult}
\alias{plotMBOResult}
\title{Plots any MBO result objects. Plots for X-Space, Y-Space and any coloumn in
the optimization path are available. This function uses
\code{\link[ParamHelpers]{plotOptPath}} from package \code{ParamHelpers}.}
\usage{
plotMBOResult(result, iters, pause = TRUE, y.over.time,
  title = "MBO Result", ...)
}
\arguments{
\item{result}{[\code{MBOResult}]\cr
\code{MBOSingleObjResult} or \code{MBOMultiObjResult} object.}

\item{iters}{[\code{integer}]\cr
Iterations to be plotted, 0 indicates the initial design.}

\item{pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{y.over.time}{[\code{list} | NULL]\cr
See argument \code{y.over.time} of the function
\code{\link[ParamHelpers]{renderOptPathPlot}} from package \code{ParamHelpers}.
Default is to plot the y variable and the infill crit in the single crit case.}

\item{title}{[\code{character(1)}]\cr
Main title for the plot.}

\item{...}{Additional parameters for the \code{\link[ParamHelpers]{plotOptPath}}
function in package \code{ParamHelpers}.}
}
\description{
Plots any MBO result objects. Plots for X-Space, Y-Space and any coloumn in
the optimization path are available. This function uses
\code{\link[ParamHelpers]{plotOptPath}} from package \code{ParamHelpers}.
}

