% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoints.R
\name{proposePoints}
\alias{proposePoints}
\title{Propose candidates for the objective function}
\usage{
proposePoints(opt.state)
}
\arguments{
\item{opt.state}{[\code{\link{OptState}}]}
}
\description{
Propose points for the objective function that should be evaluated according to the infill criterion and the recent evaluations.
}
