% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejSamp.R
\name{rejSamp}
\alias{rejSamp}
\title{Generate random param settings by rejection sampling}
\usage{
rejSamp(f, n = 1, par.set, f.max, f.min, ...)
}
\arguments{
\item{f}{[\code{function}]\cr
Function to be interpreted as probability density function.}

\item{n}{[\code{integer(1)}]\cr
Amount of random numbers to be generated. Default is 1.}

\item{par.set}{[\code{ParamSet}]\cr
ParamSet the generated random numbers are from.}

\item{...}{Arguments passed to \code[sampleValue].}
}
\value{
\code{n} Random numbers from the function \code{f} interpreted as probality density function.
}
\description{
Generate random param settings by rejection sampling
}

