% Generated by roxygen2 (4.0.1): do not edit by hand
\name{restartSavedMBO}
\alias{restartSavedMBO}
\title{Restarts a mbo-run from a save-file. Useful if your optimization is likely to
 crash, so you can start again from the iteration the crash happend and to not lose
 to much information.}
\usage{
restartSavedMBO(file)
}
\arguments{
\item{file}{[\code{character(1)}]\cr
Saved MBO run. Can be obtained by setting the \code{save.on.disk.at} argument
of the mbo control object.}
}
\description{
Restarts a mbo-run from a save-file. Useful if your optimization is likely to
 crash, so you can start again from the iteration the crash happend and to not lose
 to much information.
}

