% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMBOControlConceptDrift.R
\name{setMBOControlConceptDrift}
\alias{setMBOControlConceptDrift}
\title{Set concept drift options.}
\usage{
setMBOControlConceptDrift(control, drift.function = identity,
  window.function = identity, learn.drift = FALSE,
  calculate.th.final.point = FALSE)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{drift.function}{[\code{function}]\cr
Function that returns the position in the drift we are in, depending on the dob.}

\item{window.function}{[\code{function}]\cr
Function that returns a subset of the OptPathNg. eg. \code{function(x) {tail(x, 10)}}}

\item{learn.drift}{[\code{logical(1)}]\cr
Should the drift param be included in the surrogate?}

\item{calculate.th.final.point}{[\code{logical(1)}]\cr
Should the best predicted y be added to the opt.path after each iteration?
Note, that this will be always the theoretical best values of the previous dob!}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends an MBO control object with options for concept drift in the objective function.
}
\seealso{
Other MBOControl: \code{\link{makeMBOControl}},
  \code{\link{setMBOControlInfill}},
  \code{\link{setMBOControlMultiObj}},
  \code{\link{setMBOControlMultiPoint}},
  \code{\link{setMBOControlTermination}}
}
