% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setMBOControlMultiCrit}
\alias{setMBOControlMultiCrit}
\title{Extends mbo control object with multi-criteria specific options.}
\usage{
setMBOControlMultiCrit(control, method = "parego", parego.s,
  parego.rho = 0.05, parego.use.margin.points = rep(FALSE,
  control$number.of.targets), parego.sample.more.weights = 5L)
}
\arguments{
\item{control}{[\code{MBOControl}]\cr
MBO control object.}

\item{method}{[\code{character(1)}]\cr
Which multicrit method should be used? At the moment only parego is
supported, which is also the default.}

\item{parego.s}{[\code{integer(1)}]\cr
Parameter of parego - controls the number of weighting vectors. The default
depends on \code{number.of.targets} and leads to 100000 different possible
weight vectors. The defaults for (2, 3, 4, 5, 6) dimensions are (100000,
450, 75, 37, 23) and 10 for higher dimensions.}

\item{parego.rho}{[\code{numeric(1)}]\cr
Parameter of parego - factor for Tchebycheff function. Default 0.05 as
suggested in parego paper.}

\item{parego.sample.more.weights}{[\code{numeric(1)}]\cr
In each iteration \code{parego.sample.more.weights} * \code{propose.points}
are sampled and the weights with maximum distance to each other are chosen.
Default is 1, if only 1 point is proposed each iteration, otherwise 5.}

\item{parego.use.margin.points}{[\code{logical}]\cr
For each target function: Should the weight vector (0, ..., 0, 1, 0, ..., 0),
i.e. the weight vector with only 0 and a single 1 at the i.th position for
the i.th target function, be drawn with probability 1? Number of TRUE entries
must be less or equal to \code{propose.points}
Default is not to do this.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends mbo control object with multi-criteria specific options.
}
\note{
See the other setMBOControl... functions and \code{makeMBOControl} for referenced arguments.
}
\seealso{
makeMBOControl
}

