% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trafos}
\alias{trafoLog}
\alias{trafoSqrt}
\alias{trafos}
\title{Transformation methods.}
\format{None}
\usage{
trafoLog(base = 10)

trafoSqrt()
}
\arguments{
\item{base}{[\code{numeric(1)}]\cr
The base with respect to which logarithms are computed.
Default is \code{10}.}
}
\description{
\itemize{
  \item{\bold{logTrafo}}{\cr Natural logarithm.}
  \item{\bold{sqrtTrafo}}{\cr Square root.}
}
If negative values occur and the trafo function can handle only positive values,
a shift of the form x - min(x) + 1 is performed prior to the transformation if the
argument \code{handle.violations} is set to \dQuote{warn} which is the default
value.
}

