% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exampleRun}
\alias{exampleRun}
\title{Perform an mbo run on a test function and and visualize what happens.}
\usage{
exampleRun(fun, par.set, global.opt = NA_real_, learner, control,
  points.per.dim = 50, noisy.evals = 10, show.info = TRUE, ...)
}
\arguments{
\item{fun}{[\code{function}]\cr
See \code{\link{mbo}}.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
See \code{\link{mbo}}.}

\item{global.opt}{[\code{numeric(1)}]\cr
Objective value of global optimum, if known.
\code{NA} means not known, which is the default.
Is used to calculate and display the gap between the optimum
and the current best value in the plot.
Default is \code{NA}.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
See \code{\link{mbo}}.
Default is mlr learner \dQuote{regr.km}, which is kriging from package
DiceKriging.
\code{nugget.estim} is set to \code{TRUE} depending on whether we have
noisy observations or not.}

\item{control}{[\code{\link{MBOControl}}]\cr
See \code{\link{mbo}}.}

\item{points.per.dim}{[\code{integer}]\cr
Number of (regular spaced) locations at which to
sample the \code{fun} function per dimension.
Default is 50.}

\item{noisy.evals}{[\code{integer(1)}]\cr
Number of function evaluations per point if \code{fun} is noisy.
Default is 10.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Further arguments passed to the learner.}
}
\value{
[\code{list}]:
  \item{xseq [\code{numeric}]}{Sequence of x values from the domain of \code{fun}.}
  \item{yseq [\code{numeric}]}{Sequence of evaluated points.}
  \item{ymat [\code{numeric}]}{Sequence of evaluated points.}
  \item{mbo.res [\code{\link{MBOResult}}]}{MBO result object.}
  \item{par.set [\code{\link[ParamHelpers]{ParamSet}}]}{See argument.}
  \item{learner [\code{\link[mlr]{Learner}}]}{See argument.}
  \item{control [\code{\link{MBOControl}}]}{See argument.}
}
\description{
Usually used for 1D or 2D examples,
useful for figuring out how stuff works and for teaching purposes.
Currently only parameter spaces with numerical parameters are supported.
For visualization, run \code{plot} on the resulting object.
What is displayed is documented here: \code{\link{plot.MBOExampleRun}}.

Please note the following things:
- The true objective function (and later everything which is predicted from our surrogate model)
  is evaluated on a regular spaced grid. These evaluations are stored in the result object.
  You can control the resultion of this grid via \code{points.per.dim}.
- In every iteration the fitted, approximating surrogate model is stored in the result object
  (via \code{store.model.at} in \code{control}) so we can later visualize it quicky.
- To save a few lines of code, you can also pass a test function from package
  \code{soobench} to \code{fun}. In this case \code{par.set} and \code{global.opt}
  are autogenerated by default and the objective function works even though its
  signature does not adhere the \dQuote{parameter must be a list}-specification.
}

