% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRun.R
\name{getGlobalOpt}
\alias{getGlobalOpt}
\title{Helper function which returns the (estimated) global optimum.}
\usage{
getGlobalOpt(run)
}
\arguments{
\item{run}{[\code{MBOExampleRun}]\cr
Object of type \code{MBOExampleRun}.}
}
\value{
[\code{numeric(1)}]. (Estimated) global optimum.
}
\description{
Helper function which returns the (estimated) global optimum.
}
