% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMBOInfillCrit.R
\name{getMBOInfillCriterion}
\alias{getMBOInfillCriterion}
\alias{getMBOInfillCriterionParams}
\alias{getMBOInfillCriterionParam}
\alias{getMBOInfillCriterionName}
\alias{getMBOInfillCriterionId}
\alias{hasRequiresInfillCriterionStandardError}
\alias{getMBOInfillCriterionComponents}
\alias{getMBOInfillCritMultiplier}
\title{Get properties of MBO infill criterion.}
\usage{
getMBOInfillCriterionParams(x)

getMBOInfillCriterionParam(x, par.name)

getMBOInfillCriterionName(x)

getMBOInfillCriterionId(x)

hasRequiresInfillCriterionStandardError(x)

getMBOInfillCriterionComponents(x)

getMBOInfillCritMultiplier(x)
}
\arguments{
\item{x}{[\code{\link{MBOInfillCriterion}}]\cr
Infill criterion.}

\item{par.name}{[\code{character(1)}]\cr
Parameter name.}
}
\description{
Returns properties of an infill criterion, e.g., name or id.
}
