% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedInfillCritFunctions.R
\name{getSupportedInfillCritFunctions}
\alias{getSupportedInfillCritFunctions}
\title{Returns the names of all currently supported infill criteria.}
\usage{
getSupportedInfillCritFunctions()
}
\value{
[\code{character}]
}

