% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSupportedInfillCritFunctions}
\alias{getSupportedInfillCritFunctions}
\title{Returns the names of all currently supported infill criteria.}
\usage{
getSupportedInfillCritFunctions()
}
\value{
[\code{character}]
}
\description{
Returns the names of all currently supported infill criteria.
}

