% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSupportedInfillOptFunctions}
\alias{getSupportedInfillOptFunctions}
\title{Get names of supported infill-criteria optimizers.}
\usage{
getSupportedInfillOptFunctions()
}
\value{
[\code{character}]
}
\description{
Get names of supported infill-criteria optimizers.
}

