% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedInfillOptFunctions.R
\name{getSupportedInfillOptFunctions}
\alias{getSupportedInfillOptFunctions}
\title{Get names of supported infill-criteria optimizers.}
\usage{
getSupportedInfillOptFunctions()
}
\value{
[\code{character}]
}
\description{
None.
}
