% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSupportedMultipointInfillOptFunctions.R
\name{getSupportedMultipointInfillOptFunctions}
\alias{getSupportedMultipointInfillOptFunctions}
\title{Get names of supported multipoint infill-criteria opimizers.}
\usage{
getSupportedMultipointInfillOptFunctions()
}
\value{
[\code{character}]
}
\description{
Get names of supported multipoint infill-criteria opimizers.
}

