% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSupportedMultipointInfillOptFunctions}
\alias{getSupportedMultipointInfillOptFunctions}
\title{Get names of supported multipoint infill-criteria opimizers.}
\usage{
getSupportedMultipointInfillOptFunctions()
}
\value{
[\code{character}]
}
\description{
Get names of supported multipoint infill-criteria opimizers.
}

