% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedMultipointInfillOptFunctions.R
\name{getSupportedMultipointInfillOptFunctions}
\alias{getSupportedMultipointInfillOptFunctions}
\title{Get names of supported multipoint infill-criteria optimizers.}
\usage{
getSupportedMultipointInfillOptFunctions()
}
\value{
[\code{character}]
}
\description{
Returns all names of supported multipoint infill-criteria optimizers.
}
