% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infill_crits.R
\name{infillcrits}
\alias{infillcrits}
\alias{makeMBOInfillCriterionAEI}
\alias{makeMBOInfillCriterionCB}
\alias{makeMBOInfillCriterionDIB}
\alias{makeMBOInfillCriterionEI}
\alias{makeMBOInfillCriterionEQI}
\alias{makeMBOInfillCriterionMeanResponse}
\alias{makeMBOInfillCriterionStandardError}
\title{Infill criteria.}
\usage{
makeMBOInfillCriterionMeanResponse()

makeMBOInfillCriterionStandardError()

makeMBOInfillCriterionEI(se.threshold = 1e-06)

makeMBOInfillCriterionCB(cb.lambda = 1, cb.inflate.se = FALSE)

makeMBOInfillCriterionAEI(aei.use.nugget = FALSE, se.threshold = 1e-06)

makeMBOInfillCriterionEQI(eqi.beta = 0.75, se.threshold = 1e-06)

makeMBOInfillCriterionDIB(cb.lambda = 1, cb.inflate.se = FALSE)
}
\arguments{
\item{se.threshold}{[\code{numeric(1)}]\cr
In order to avoid numerical problems the standard error estimation is assumed to
be exactly zero, if it is below \code{se.threshold}.
Default is 1e-6.}

\item{cb.lambda}{[\code{numeric(1)}]\cr
Lambda parameter for confidence bound infill criterion.
In the multi-objective case we recommend to set this value to
\eqn{q(0.5 \cdot \pi_{CB}^(1 / n))} where \eqn{q} is the quantile
function of the standard normal distribution, \eqn{\pi_CB} is the probability
of improvement value and \eqn{n} is the number of objectives of the considered problem.
Default is 1.}

\item{cb.inflate.se}{[\code{logical(1)}]\cr
Try to inflate or deflate the estimated standard error to get to the same scale as the mean?
Calculates the range of the mean and standard error and multiplies the standard error
with the quotient of theses ranges.
Default is \code{FALSE}.}

\item{aei.use.nugget}{[\code{logical(1)}]\cr
Should the nugget effect be used for the pure variance estimation for augmented
expected improvement?
Default is \code{FALSE}.}

\item{eqi.beta}{[\code{numeric(1)}]\cr
Beta parameter for expected quantile improvement criterion.
Default is 0.75.}
}
\description{
\pkg{mlrMBO} contains most of the most popular infill criteria, e.g., expected
improvement, (lower) confidence bound etc. Moreover, custom infill criteria
may be generated with the \code{\link{makeMBOInfillCriterion}} function.
}

