% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infill_crits.R, R/zzz.R
\name{infillcrits}
\alias{infillcrits}
\alias{makeMBOInfillCriterionMeanResponse}
\alias{makeMBOInfillCriterionStandardError}
\alias{makeMBOInfillCriterionEI}
\alias{makeMBOInfillCriterionCB}
\alias{makeMBOInfillCriterionAEI}
\alias{makeMBOInfillCriterionEQI}
\alias{makeMBOInfillCriterionDIB}
\alias{crit.ei}
\title{Infill criteria.}
\usage{
makeMBOInfillCriterionMeanResponse()

makeMBOInfillCriterionStandardError()

makeMBOInfillCriterionEI(se.threshold = 1e-06)

makeMBOInfillCriterionCB(cb.lambda = NULL)

makeMBOInfillCriterionAEI(aei.use.nugget = FALSE, se.threshold = 1e-06)

makeMBOInfillCriterionEQI(eqi.beta = 0.75, se.threshold = 1e-06)

makeMBOInfillCriterionDIB(cb.lambda = 1, sms.eps = NULL)
}
\arguments{
\item{se.threshold}{[\code{numeric(1)}]\cr
In order to avoid numerical problems the standard error estimation is assumed to
be exactly zero, if it is below \code{se.threshold}.
Default is 1e-6.}

\item{cb.lambda}{[\code{numeric(1)}]\cr
Lambda parameter for confidence bound infill criterion.
In the multi-objective case we recommend to set this value to
\eqn{q(0.5 \cdot \pi_{CB}^(1 / n))} where \eqn{q} is the quantile
function of the standard normal distribution, \eqn{\pi_CB} is the probability
of improvement value and \eqn{n} is the number of objectives of the considered problem.
Default is 1 in case of a fully numeric parameter set and 2 otherwise.}

\item{aei.use.nugget}{[\code{logical(1)}]\cr
Should the nugget effect be used for the pure variance estimation for augmented
expected improvement?
Default is \code{FALSE}.}

\item{eqi.beta}{[\code{numeric(1)}]\cr
Beta parameter for expected quantile improvement criterion.
Default is 0.75.}

\item{sms.eps}{[\code{numeric(1)} | \code{NULL}]\cr
Epsilon for epsilon-dominance for \code{dib.indicator = "sms"}.
Default is \code{NULL}, in this case it is adaptively set.}
}
\description{
\pkg{mlrMBO} contains most of the most popular infill criteria, e.g., expected
improvement, (lower) confidence bound etc. Moreover, custom infill criteria
may be generated with the \code{\link{makeMBOInfillCriterion}} function.
}
\section{Predefined standard Infill Criteria}{

For commonly used settings of the Infill criteria:
\describe{
\item{crit.ei}{Expected Improvement}
\item{crit.mr}{Mean Response}
\item{crit.se}{Standard Error}
\item{crit.cb}{Confidence Bound with Lambda dependent on Parameter Space}
\item{crit.cb1}{Confidence Bound with Lambda=1}
\item{crit.cb2}{Confidence Bound with Lambda=2}
\item{crit.cb3}{Confidence Bound with Lambda=3}
\item{crit.cb4}{Confidence Bound with Lambda=4}
\item{crit.cb5}{Confidence Bound with Lambda=5}
\item{crit.aei}{Augmeted expected improvement}
\item{crit.eqi}{Expected quantile improvement}
\item{crit.dib1}{Directed Indicator Based Search with Lambda=1}
\item{crit.dib2}{Directed Indicator Based Search with Lambda=2}
\item{crit.dib3}{Directed Indicator Based Search with Lambda=3}
\item{crit.dib4}{Directed Indicator Based Search with Lambda=4}
\item{crit.dib5}{Directed Indicator Based Search with Lambda=5}

\item{cv10}{10-fold cross-validation}
}
}

