% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMBOTrafoFunction.R
\name{makeMBOTrafoFunction}
\alias{makeMBOTrafoFunction}
\title{Create a transformation function for MBOExampleRun.}
\usage{
makeMBOTrafoFunction(name, fun)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of the transformation.}

\item{fun}{[\code{function}]\cr
R function which expects a numeric vector.}
}
\value{
Object of type MBOTrafoFunction.
}
\description{
Creates a transformation function for MBOExampleRun.
}
\seealso{
\link{trafos}
}
