% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mbo}
\alias{mbo}
\title{Optimizes a function with sequential model based optimization.}
\usage{
mbo(fun, par.set, design = NULL, learner, control, show.info = TRUE,
  more.args = list())
}
\arguments{
\item{fun}{[\code{function(x, ...)}]\cr
Fitness function to minimize. The first argument has to be a list of values.
The function has to return a single numerical value.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
See \code{\link{mbo}}.}

\item{design}{[\code{data.frame} | NULL]\cr
Initial design as data frame.
If the parameters have corresponding trafo functions,
the design must not be transformed before it is passed!
If \code{NULL}, one is constructed from the settings in \code{control}.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
Regression learner to model \code{fun}.}

\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}

\item{more.args}{[list]\cr
Further arguments passed to fitness function.}
}
\value{
[\code{\link{MBOSingleObjResult}} | \code{\link{MBOMultiObjResult}}]
}
\description{
Optimizes a function with sequential model based optimization.
}
\note{
It is possible to parallelize the evaluation of the target function to speed up the computation.
Internally the evaluation of the target function is realized with the R package parallelMap. See the mlrMBO tutorial
respectively the help pages of \code{\link[parallelMap]{parallelMap}} for instructions on how to set up parallization.
}

