% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo.R
\name{mbo}
\alias{mbo}
\title{Optimizes a function with sequential model based optimization.}
\usage{
mbo(fun, design = NULL, learner = NULL, control,
  show.info = getOption("mlrMBO.show.info", TRUE), more.args = list())
}
\arguments{
\item{fun}{[\code{function(x, ...)}]\cr
Fitness function to minimize. The first argument has to be a list of values.
The function has to return a single numerical value.
In fact it is possible to return even more information which will be stored
in the optimization path. To achieve this, simply append the attribute \dQuote{extras}
to the return value of the target function. This has to be a named list of scalar values.
Each of these values will be stored additionally in the optimization path.}

\item{design}{[\code{data.frame}]\cr
Initial design as data frame. If the y-values are not already present in design,
mbo will evaluate the points.
If the parameters have corresponding trafo functions,
the design must not be transformed before it is passed!
Functions to generate designs are available in \code{ParamHelpers}: \code{\link[ParamHelpers]{generateDesign}}, \code{\link[ParamHelpers]{generateGridDesign}}, \code{\link[ParamHelpers]{generateRandomDesign}}.
Default is \code{NULL}, which means \code{\link[ParamHelpers]{generateDesign}} is called and a design
of size 4 times number of all parameters is created. The points are drawn via \code{\link[lhs]{maximinLHS}} to maximize the minimal distance between design points.}

\item{learner}{[\code{\link[mlr]{Learner}}]\cr
Regression learner from mlr, which is used as a surrogate to model our fitness function.
If \code{NULL} (default), the default learner is determined as described here: \link{mbo_default_learner}.}

\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}

\item{more.args}{[list]\cr
Further arguments passed to fitness function.}
}
\value{
[\code{\link{MBOSingleObjResult}} | \code{\link{MBOMultiObjResult}}]
}
\description{
See \link{mbo_parallel} for all parallelization options.
}
\examples{
obj.fun = makeSingleObjectiveFunction(
 fn = function(x) x[1]^2 + sin(x[2]),
 par.set = makeNumericParamSet(id = "x", lower = -1, upper = 1, len = 2))
ctrl = makeMBOControl()
ctrl = setMBOControlTermination(ctrl, iters = 3L)
des = generateDesign(n = 5L, getParamSet(obj.fun), fun = lhs::maximinLHS)
res = mbo(obj.fun, design = des, control = ctrl)
print(res)
plot(res)
}

