% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboContinue.R
\name{mboContinue}
\alias{mboContinue}
\title{Continues an mbo run from a save-file.}
\usage{
mboContinue(opt.state)
}
\arguments{
\item{opt.state}{[\code{OptState} | \code{character(1)}]\cr
Either the \code{save.state} slot of an \code{MBOResult} object, or a file path of a saved MBO state.
See \code{save.on.disk.at} argument of \code{\link{MBOControl}} object.}
}
\value{
See \code{\link{mbo}}.
}
\description{
Useful if your optimization is likely to crash,
so you can continue from a save point and will not lose too much information and runtime.
}
