% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MBOMultiObjResult}
\alias{MBOMultiObjResult}
\title{Multi-Objective result object.}
\description{
\itemize{
  \item{pareto.front [\code{matrix}]}{Pareto front of all evaluated points.}
  \item{pareto.set [\code{list} of \code{list}s]}{Pareto set of all evaluated points.}
  \item{opt.path [\code{\link[ParamHelpers]{OptPath}}]}{Optimization path.}
  \item{models [List of \code{\link[mlr]{WrappedModel}}]}{List of saved regression models.}
}
}

