% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MBOMultiObjResult}
\alias{MBOMultiObjResult}
\title{Multi-Objective result object.}
\description{
\itemize{
  \item{pareto.front [\code{matrix}]}{Pareto front of all evaluated points.}
  \item{pareto.set [\code{list} of \code{list}s]}{Pareto set of all evaluated points.}
  \item{pareto.inds [\code{numeric}]}{Indizes of the Pareto-optimal points in the opt.path}
  \item{opt.path [\code{\link[ParamHelpers]{OptPath}}]}{Optimization path.
    Includes all evaluated points and additional information.
    You can convert it via \code{as.data.frame}.}
  \item{models [List of \code{\link[mlr]{WrappedModel}}]}{List of saved regression models.}
}
}

