% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OptState.R
\name{plot.OptState}
\alias{plot.OptState}
\title{Generate ggplot2 Object}
\usage{
\method{plot}{OptState}(x, scale.panels = FALSE, ...)
}
\arguments{
\item{x}{[\code{OptState}]\cr
The OptState.}

\item{scale.panels}{[\code{logical(1)}]\cr
If \code{TRUE} the values in each panel will be scaled to [0,1].}

\item{...}{[any] \cr
Not used.}
}
\description{
Plots the values of the infill criterion for a 1- and 2-dimensional numerical search space for a given \code{\link{OptState}}.
}
