% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MBOResult.R
\name{plotMBOResult}
\alias{plotMBOResult}
\alias{plot.MBOSingleObjResult}
\alias{plot.MBOMultiObjResult}
\title{MBO Result Plotting}
\usage{
\method{plot}{MBOSingleObjResult}(x, iters = NULL, pause = interactive(), ...)

\method{plot}{MBOMultiObjResult}(x, iters = NULL, pause = interactive(), ...)
}
\arguments{
\item{x}{[\code{MBOResult}]\cr
\code{MBOSingleObjResult} or \code{MBOMultiObjResult} object.}

\item{iters}{[\code{integer}]\cr
Iterations to be plotted, 0 indicates the initial design. Default is all iterations.}

\item{pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{interactive()}.}

\item{...}{Additional parameters for the \code{\link[ParamHelpers]{plotOptPath}}
function in package \code{ParamHelpers}.}
}
\description{
Plots any MBO result objects. Plots for X-Space, Y-Space and any column in
the optimization path are available. This function uses
\code{\link[ParamHelpers]{plotOptPath}} from package \code{ParamHelpers}.
}
