% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMBOControl.R
\name{print.MBOControl}
\alias{print.MBOControl}
\title{Print mbo control object.}
\usage{
\method{print}{MBOControl}(x, ...)
}
\arguments{
\item{x}{[\code{\link{MBOControl}}]\cr
Control object.}

\item{...}{[any]\cr
Not used.}
}
\description{
Print mbo control object.
}
