% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMBOControlConceptDrift.R
\name{setMBOControlConceptDrift}
\alias{setMBOControlConceptDrift}
\title{Set concept drift options.}
\usage{
setMBOControlConceptDrift(control, drift.param = NULL,
  drift.function = NULL, window.function = NULL)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{drift.param}{[\code{character(1)}]\cr
Which parameter determines the drift position we are in?}

\item{drift.function}{[\code{function}]\cr
Function that returns the position in the drift we are in, depending on the dob.}

\item{window.function}{[\code{function}]\cr
Function that returns a subset of the OptPathNg. eg. \code{function(op) {r = op$clone(); r$data = tail(r$data); r}}}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends an MBO control object with options for concept drift in the objective function.
}
\seealso{
Other MBOControl: \code{\link{makeMBOControl}},
  \code{\link{setMBOControlInfill}},
  \code{\link{setMBOControlMultiObj}},
  \code{\link{setMBOControlMultiPoint}},
  \code{\link{setMBOControlTermination}}
}
