% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setMBOControlInfill}
\alias{setMBOControlInfill}
\title{Extends mbo control object with infill criteria and infill optimizer options.}
\usage{
setMBOControlInfill(control, crit = "mean", crit.lcb.lambda = 1,
  opt = "focussearch", opt.restarts = 1L, opt.focussearch.maxit = 5L,
  opt.focussearch.points = 10000L, opt.cmaes.control = list(),
  opt.ea.maxit = 500L, opt.ea.mu = 10L, opt.ea.sbx.eta = 15,
  opt.ea.sbx.p = 0.5, opt.ea.pm.eta = 15, opt.ea.pm.p = 0.5,
  opt.ea.lambda = 1L)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
MBO control object.}

\item{crit}{[\code{character(1)}]\cr
How should infill points be rated. Possible parameter values are:
\dQuote{mean}: Mean response.
\dQuote{ei}: Expected improvement.
\dQuote{aei}: Augmented expected improvement.
\dQuote{lcb}: Lower confidence bound.
\dQuote{multiFid}: Multifidelity: Expected improvement on different levels of the perfomance parameter defined in the \code{MBOMultiFidControl}.
Alternatively, you may pass a function name as string.}

\item{crit.lcb.lambda}{[\code{numeric(1)}]\cr
Lambda parameter for lower confidence bound infill criterion.
Only used if \code{crit == "lcb"}, ignored otherwise.
Default is 1.}

\item{opt}{[\code{character(1)}]\cr
How should SINGLE points be proposed by using the surrogate model. Possible values are:
\dQuote{focussearch}: In several iteration steps the parameter space is
focused on an especial promising region according to infill criterion.
\dQuote{cmaes}: Use CMAES to optimize infill criterion. If all CMAES runs fail, a random point is generated
instead and a warning informs about it.
\dQuote{ea}: Use an (mu+1) EA to optimize infill criterion.
Default is \dQuote{focussearch}.
Alternatively, you may pass a function name as string.}

\item{opt.restarts}{[\code{integer(1)}]\cr
Number of independent restarts for optimizer of infill criterion.
If \code{opt == "cmaes"} the first start point for the optimizer is always the
currently best point in the design of already visited points.
Subsequent restarts are started at random points.
Default is 1.}

\item{opt.focussearch.maxit}{[\code{integer(1)}]\cr
For \code{opt = "focussearch"}:
Number of iteration to shrink local focus.
Default is 5.}

\item{opt.focussearch.points}{[\code{integer(1)}]\cr
For \code{opt = "focussearch"}:
Number of points in each iteration of the focus search optimizer.
Default is 10000.}

\item{opt.cmaes.control}{[\code{list}]\cr
For \code{opt = "cmaes"}:
Control argument for cmaes optimizer.
Default is empty list.}

\item{opt.ea.maxit}{[\code{integer(1)}]\cr
For \code{opt = "ea"}:
Number of iterations / generations of EA.
Default is 500.}

\item{opt.ea.mu}{[\code{integer(1)}]\cr
For \code{opt = "ea"}:
Population size of EA.
Default is 10.}

\item{opt.ea.pm.eta}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Distance parameter of mutation distribution, see \code{\link[emoa]{pm_operator}}.
Default is 15.}

\item{opt.ea.pm.p}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Probability of 1-point mutation, see \code{\link[emoa]{pm_operator}}.
Default is 0.5.}

\item{opt.ea.sbx.eta}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Distance parameter of crossover distribution , see \code{\link[emoa]{sbx_operator}}.
Default is 15.}

\item{opt.ea.sbx.p}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Probability of 1-point crossover, see \code{\link[emoa]{sbx_operator}}.
Default is 0.5.}

\item{opt.ea.lambda}{[\code{numeric{1}}]\cr
For \code{opt.ea = "ea"}.
Number of children generated in each generation.
Default is 1.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends mbo control object with infill criteria and infill optimizer options.
}
\seealso{
makeMBOControl
}

