% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setMBOControlInfill}
\alias{setMBOControlInfill}
\title{Extends mbo control object with infill criteria and infill optimizer options.}
\usage{
setMBOControlInfill(control, crit = NULL, crit.eqi.beta = 0.75,
  crit.lcb.lambda = 1, crit.lcb.pi = NULL, filter.proposed.points = NULL,
  filter.proposed.points.tol = NULL, opt = "focussearch",
  opt.restarts = NULL, opt.focussearch.maxit = NULL,
  opt.focussearch.points = NULL, opt.cmaes.control = NULL,
  opt.ea.maxit = NULL, opt.ea.mu = NULL, opt.ea.sbx.eta = NULL,
  opt.ea.sbx.p = NULL, opt.ea.pm.eta = NULL, opt.ea.pm.p = NULL,
  opt.ea.lambda = NULL, opt.nsga2.popsize = NULL,
  opt.nsga2.generations = NULL, opt.nsga2.cprob = NULL,
  opt.nsga2.cdist = NULL, opt.nsga2.mprob = NULL, opt.nsga2.mdist = NULL)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{crit}{[\code{character(1)}]\cr
How should infill points be rated. Possible parameter values are:
\dQuote{mean}: Mean response.
\dQuote{ei}: Expected improvement.
\dQuote{aei}: Augmented expected improvement.
\dQuote{eqi}: Expected quantile improvement.
\dQuote{lcb}: Lower confidence bound.
\dQuote{multiFid}: Multifidelity: Expected improvement on different levels of the perfomance parameter defined in the \code{MBOMultiFidControl}.
Alternatively, you may pass a function name as string.}

\item{crit.eqi.beta}{[\code{numeric(1)}]\cr
Beta parameter for expected quantile improvement criterion.
Only used if \code{crit == "eqi"}, ignored otherwise.
Default is 0.75.}

\item{crit.lcb.lambda}{[\code{numeric(1)}]\cr
Lambda parameter for lower confidence bound infill criterion.
Only used if \code{crit == "lcb"}, ignored otherwise.
Default is 1.}

\item{crit.lcb.pi}{[\code{numeric(1)}]\cr
Probability-of-improvement value to determine the lambda parameter for lcb infill criterion.
It is an alternative to set the trade-off between \dQuote{mean} and \dQuote{se}.
Only used if \code{crit == "lcb"}, ignored otherwise.
If specified, \code{crit.lcb.lambda == NULL} must hold.
Default is \code{NULL}.}

\item{filter.proposed.points}{[\code{logical(1)}]\cr
Design points located too close to each other can lead to
numerical problems when using e.g. kriging as a surrogate model.
This parameter activates or deactivates a heuristic to handle this issue.
If \code{TRUE}, proposed points whose distance to design points or other current
candidate points is smaller than \code{filter.proposed.points.tol}, are replaced by random points.
If enabled, a logical column \dQuote{filter.replace} is added to the resulting \code{opt.path},
so you can see whether such a replacement happened.
Default is \code{FALSE}.}

\item{filter.proposed.points.tol}{[\code{numeric(1)}]\cr
Tolerance value filtering of proposed points. We currently use a maximum metric
to calculate the distance between points.
Default is 0.0001.}

\item{opt}{[\code{character(1)}]\cr
How should SINGLE points be proposed by using the surrogate model. Possible values are:
\dQuote{focussearch}: In several iteration steps the parameter space is
focused on an especial promising region according to infill criterion.
\dQuote{cmaes}: Use CMAES to optimize infill criterion. If all CMAES runs fail, a random point is generated
instead and a warning informs about it.
\dQuote{ea}: Use an (mu+1) EA to optimize infill criterion.
\dQuote{nsga2}: NSGA2 for multi obj. optimizationen. Needed for mspot.
Default is \dQuote{focussearch}.
Alternatively, you may pass a function name as string.}

\item{opt.restarts}{[\code{integer(1)}]\cr
Number of independent restarts for optimizer of infill criterion.
If \code{opt == "cmaes"} the first start point for the optimizer is always the
currently best point in the design of already visited points.
Subsequent restarts are started at random points.
Default is 1.}

\item{opt.focussearch.maxit}{[\code{integer(1)}]\cr
For \code{opt = "focussearch"}:
Number of iteration to shrink local focus.
Default is 5.}

\item{opt.focussearch.points}{[\code{integer(1)}]\cr
For \code{opt = "focussearch"}:
Number of points in each iteration of the focus search optimizer.
Default is 10000.}

\item{opt.cmaes.control}{[\code{list}]\cr
For \code{opt = "cmaes"}:
Control argument for cmaes optimizer.
Default is empty list.}

\item{opt.ea.maxit}{[\code{integer(1)}]\cr
For \code{opt = "ea"}:
Number of iterations / generations of EA.
Default is 500.}

\item{opt.ea.mu}{[\code{integer(1)}]\cr
For \code{opt = "ea"}:
Population size of EA.
Default is 10.}

\item{opt.ea.sbx.eta}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Distance parameter of crossover distribution , see \code{\link[emoa]{sbx_operator}}.
Default is 15.}

\item{opt.ea.sbx.p}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Probability of 1-point crossover, see \code{\link[emoa]{sbx_operator}}.
Default is 0.5.}

\item{opt.ea.pm.eta}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Distance parameter of mutation distribution, see \code{\link[emoa]{pm_operator}}.
Default is 15.}

\item{opt.ea.pm.p}{[\code{numeric(1)}]\cr
For \code{opt = "ea"}:
Probability of 1-point mutation, see \code{\link[emoa]{pm_operator}}.
Default is 0.5.}

\item{opt.ea.lambda}{[\code{numeric{1}}]\cr
For \code{opt.ea = "ea"}.
Number of children generated in each generation.
Default is 1.}

\item{opt.nsga2.popsize}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
Population size of nsga2.
Default is 100.}

\item{opt.nsga2.generations}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
Number of populations for of nsga2.
Default is 50.}

\item{opt.nsga2.cprob}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
nsga2 param. Default is 0.7.}

\item{opt.nsga2.cdist}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
nsga2 param. Default is 5.}

\item{opt.nsga2.mprob}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
nsga2 param. Default is 0.2.}

\item{opt.nsga2.mdist}{[\code{numeric{1}}]\cr
For \code{opt.multicrit.method = "nsga2"}.
nsga2 param. Default is 10.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends mbo control object with infill criteria and infill optimizer options.
}
\note{
See the other setMBOControl... functions and \code{makeMBOControl} for referenced arguments.
}
\seealso{
makeMBOControl
}

