% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMBOControlMultiFid.R
\name{setMBOControlMultiFid}
\alias{setMBOControlMultiFid}
\title{Set multi-fidelity options.}
\usage{
setMBOControlMultiFid(control, param, lvls, costs = NULL,
  cor.grid.points = NULL, force.last.level.steps = NULL, show.info = NULL)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{param}{[\code{character(1)}]\cr
The name of the parameter which increases the performance but also calculation costs.}

\item{lvls}{[\code{numeric}]\cr
The values of the param the learner should be trained with, in ascending order.}

\item{costs}{[\code{numeric}]\cr
Vector defining the cost for each level.
Default is \code{NULL} which means that the cost will be predicted by a model build on the \code{exec.time} from the so far evaluated points.}

\item{cor.grid.points}{[\code{integer(1)}]\cr
Numbers of points used to calculate the correlation between the different levels of
the \code{param}.}

\item{force.last.level.steps}{[\code{integer(1)}]
Force an evaluation on the highest fidelity level each i-th step.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends an MBO control object with multi-fidelity specific options.
}
\seealso{
Other MBOControl: \code{\link{makeMBOControl}},
  \code{\link{setMBOControlInfill}},
  \code{\link{setMBOControlMultiObj}},
  \code{\link{setMBOControlMultiPoint}},
  \code{\link{setMBOControlTermination}}
}

