% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setMBOControlMultiFid.R
\name{setMBOControlMultiFid}
\alias{setMBOControlMultiFid}
\title{Extends mbo control object with multiFid-algorithm specific options.}
\usage{
setMBOControlMultiFid(control, param, lvls, costs = NULL,
  cor.grid.points = NULL, force.last.level.evals = NULL,
  eval.lower = NULL, show.info = NULL)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{param}{[\code{character(1)}]\cr
The name of the parameter which increases the performance but also calculation costs.}

\item{lvls}{[\code{numeric}]\cr
The values of the param the learner should be trained with, in ascending order.}

\item{costs}{[\code{numeric}]\cr
Vector defining the cost for each level.
Default is \code{NULL} which means that the cost will be predicted by a model build on the \code{exec.time} from the so far evaluated points.}

\item{cor.grid.points}{[\code{integer(1)}]\cr
Numbers of points used to calculate the correlation between the different levels of
the \code{param}.}

\item{force.last.level.evals}{[\code{integer(1)}]
How many evaluations should be done on the last value of fid.param?}

\item{eval.lower}{[\code{boolean(1)}]\cr
For each evaluation, evaluate the lower fidelity levels as well. Default is \code{FALSE}.}

\item{show.info}{[\code{logical(1)}]\cr
Verbose output on console?
Default is \code{TRUE}.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends mbo control object with multiFid-algorithm specific options.
}
\note{
See the other setMBOControl... functions and \code{makeMBOControl} for referenced arguments.
}
\seealso{
makeMBOControl
}

