% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setMBOControlMultiFid}
\alias{setMBOControlMultiFid}
\title{Extends mbo control object with multiFid-algorithm specific options.}
\usage{
setMBOControlMultiFid(control, param, lvls, costs = NULL,
  cor.grid.points = NULL, force.last.level.evals = NULL,
  eval.lower = NULL, show.info = NULL)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{param}{[\code{character(1)}]\cr
The name of the parameter which increases the performance but also calculation costs.
Must be a discrete parameter.}

\item{lvls}{[\code{numeric}]\cr
The values of the param the learner should be trained with.}

\item{costs}{[\code{function}]\cr
Vektorized (?) cost function with the params \code{cur} and \code{last}.}

\item{cor.grid.points}{[\code{integer(1)}]\cr
Numbers of points used to calculate the correlation between the different levels of
the \code{param}.}

\item{force.last.level.evals}{[\code{integer(1)}]
How many evaluations should be done on the last value of fid.param?}

\item{eval.lower}{[\code{boolean(1)}]\cr
For each evaluation, evaluate the lower fidelity levels as well. Default is \code{FALSE}.}
}
\value{
[\code{\link{MBOControl}}].
}
\description{
Extends mbo control object with multiFid-algorithm specific options.
}
\note{
See the other setMBOControl... functions and \code{makeMBOControl} for referenced arguments.
}
\seealso{
makeMBOControl
}

