% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMBOControlRefinement.R
\name{setMBOControlRefinement}
\alias{setMBOControlRefinement}
\title{Sets Refinement Control}
\usage{
setMBOControlRefinement(control, refinement.control, refinement.learner)
}
\arguments{
\item{control}{[\code{\link{MBOControl}}]\cr
Control object for mbo.}

\item{refinement.control}{[\code{MBOControl}]}

\item{refinement.learner}{[\code{\link[mlr]{Learner}}]}
}
\description{
Put another control object into the mbo control object to control the refinement.
}
