% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trafos}
\alias{log10Trafo}
\alias{logTrafo}
\alias{sqrtTrafo}
\alias{trafos}
\title{Transformation methods.}
\format{None}
\usage{
logTrafo()

log10Trafo()

sqrtTrafo()
}
\description{
\itemize{
\item{\bold{logTrafo}}{\cr Natural logarithm.}
\item{\bold{log10Trafo}}{\cr Logartihm to base 10.}
\item{\bold{sqrtTrafo}}{\cr Square root.}
}
If negative values occur and the trafo function can handle only positive values,
a shift of the form x - min(x) + 1 is performed prior to the transformation.
}

