% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMBOTrafoFunction.R
\name{trafos}
\alias{trafos}
\alias{trafoLog}
\alias{trafoSqrt}
\title{Transformation methods.}
\format{None}
\usage{
trafoLog(base = 10, handle.violations = "warn")

trafoSqrt(handle.violations = "warn")
}
\arguments{
\item{base}{[\code{numeric(1)}]\cr
The base with respect to which logarithms are computed.
Default is \code{10}.}

\item{handle.violations}{[\code{character(1)}]\cr
What should be done, if negative values occur? Setting this option to
\dQuote{warn}, which is the default, shifts the function. \dQuote{error}
stops the process immediately.}
}
\description{
\itemize{
  \item{\bold{logTrafo}}{\cr Natural logarithm.}
  \item{\bold{sqrtTrafo}}{\cr Square root.}
}
If negative values occur and the trafo function can handle only positive values,
a shift of the form x - min(x) + 1 is performed prior to the transformation if the
argument \code{handle.violations} is set to \dQuote{warn} which is the default
value.
}
