% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMBO.R
\name{updateSMBO}
\alias{updateSMBO}
\title{Updates SMBO with the new observations}
\usage{
updateSMBO(opt.state, x, y)
}
\arguments{
\item{opt.state}{[\code{\link{OptState}}]
The optimization state.
Generated by \code{\link{initSMBO}}, this function or an \code{\link{mbo}} run.}

\item{x}{[\code{data.frame}]
Named x values.
One row per set of x values.}

\item{y}{[\code{numeric|list}]
Outcome of the optimization.
For multiple results use a list.
For a result of a multi-objective function use a numeric vector.}
}
\value{
[\code{\link{OptState}}]
}
\description{
After a function evaluation you want to update the \code{\link{OptState}} to get new proposals.
}
