% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapSmoofConceptDrift.R
\name{wrapSmoofConceptDrift}
\alias{wrapSmoofConceptDrift}
\title{Wrap a smoof function for concept drift.}
\usage{
wrapSmoofConceptDrift(fn, drift.param)
}
\arguments{
\item{fn}{[\code{smoof_function}]\cr
The smoof function that you want to wrap.}

\item{drift.param}{[\code{character(1)}]\cr
The parameter that defines the drift.}
}
\value{
return_type
}
\description{
Wrap a smoof function for concept drift.
}
